pyproject.toml
[project]
name = "cv-match-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langdetect>=1.0.9",
    "numpy<2.0.0",
    "protobuf>=3.20.0",
    "pypdf2>=3.0.1",
    "pytest>=8.3.5",
    "python-docx>=1.1.0",
    "python-dotenv>=1.1.0",
    "transformers>=4.52.4",
    "torch>=2.0.0",
    "tiktoken>=0.9.0",
    "docx>=0.2.4",
]

[tool.pytest]
pythonpath = [
    "."
]
